feat: track debounced jobs as separate metrics (Laravel 13.6+)
* feat: add JobMetricsDebounced event for downstream consumers


* feat: add DebouncedJobTracker for inter-listener communication


* feat: add recordDebounced() to JobMetricsRepository contract and implementations


* feat: add RecordJobDebouncedAction


* feat: add totalDebounced to JobExecutionData DTO


* feat: add JobDebouncedListener for debounce metrics tracking


* feat: conditionally register JobDebouncedListener for Laravel 13.6+


* feat: skip performance metrics for debounced jobs in JobProcessedListener


* feat: export job_debounced_total counter to Prometheus

* fix: resolve PHPStan level 9 errors

- Add @phpstan-ignore for object property access in JobDebouncedListener
  (event type only exists in Laravel 13.6+)
- Make total_debounced optional in Prometheus PHPDoc type hint
- Remove redundant null coalesce on non-nullable WorkerHeartbeat::$hostname


* fix: cast memory division to float in ChildProcessTrackingTest

Integer division can return int when result is whole number,
causing toBeFloat() assertion to fail.


---------

+++ tests/Feature/ChildProcessTrackingTest.php
@@ -45,7 +45,7 @@ it('tracks child processes when job spawns subprocesses', function () {
     $stats = $statsResult->getValue();
 
     // Verify metrics are calculated correctly as in JobProcessedListener
-    $memoryMb = $stats->peak->memoryRssBytes / 1024 / 1024;
+    $memoryMb = (float) ($stats->peak->memoryRssBytes / 1024 / 1024);
     expect($memoryMb)->toBeGreaterThan(0.0);
 
     $cpuUsagePercent = $stats->delta->cpuUsagePercentage();
@@ -160,7 +160,7 @@ it('provides process resource usage compatible with JobProcessedListener calcula
     $stats = $statsResult->getValue();
 
     // This is exactly how JobProcessedListener uses the metrics:
-    $memoryMb = $stats->peak->memoryRssBytes / 1024 / 1024;
+    $memoryMb = (float) ($stats->peak->memoryRssBytes / 1024 / 1024);
     $cpuUsagePercent = $stats->delta->cpuUsagePercentage();
     $durationSeconds = $stats->delta->durationSeconds;
     $cpuTimeMs = ($cpuUsagePercent / 100.0) * $durationSeconds * 1000.0;
